\name{astro}
\alias{astro}
\title{Astronomical arguments and speeds}
\usage{astro(t)}
\arguments{
\item{t}{POSIXct time in UTC}
}
\value{A named list of astronomical values and speeds (deg and deg/hour).}
\description{Computes astronomical arguments and their speeds used in harmonic tidal analysis.}
